requirements.txt
flask==2.3.3
Jinja2==3.1.2
MarkupSafe==2.1.3
pip==23.2.1
python-dotenv==1.0.0
requests==2.26.0